refactor(tech): 코드리뷰 반영

+++ src/main/java/com/dreamypatisiel/devdevdev/elastic/data/domain/ElasticSlice.java
@@ -1,16 +1,15 @@
 package com.dreamypatisiel.devdevdev.elastic.data.domain;
 
+import java.util.List;
 import lombok.Getter;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.SliceImpl;
 
-import java.util.List;
-
 @Getter
 public class ElasticSlice<T> extends SliceImpl<T> {
-    private final Long totalElements;
+    private final int totalElements;
 
-    public ElasticSlice(List<T> content, Pageable pageable, Long totalElements, boolean hasNext) {
+    public ElasticSlice(List<T> content, Pageable pageable, int totalElements, boolean hasNext) {
         super(content, pageable, hasNext);
         this.totalElements = totalElements;
     }
